Simplfy geometry creation, and reduce dll size
git-svn-id: https://nekonme.googlecode.com/svn/trunk@107 1509560c-5e2a-0410-865c-31c25e1cfdef

+++ project/renderer/Utils.cpp
@@ -2,8 +2,7 @@
 #include "AA.h"
 
 Uint8  AA4x::mDrawing[32];
-Uint8  AA4x::mAlpha[32];
-
+int  AA4x::mAlpha[32];
 
 
 // Find y-extent of object, this is in pixels, and is the intersection
